+++ setup.py
@@ -22,7 +22,8 @@ setup(name='tag',
       author_email='daniel.standage@gmail.com',
       license='BSD-3',
       packages=['tag'],
-      scripts=list(glob.glob('scripts/*.py')),
+      scripts=list(glob.glob('bin/tag')),
+      install_requires=['intervaltree'],
       classifiers=[
           'Development Status :: 4 - Beta',
           'Environment :: Console',
