+++ tag/cli/__init__.py
@@ -8,4 +8,4 @@
 # -----------------------------------------------------------------------------
 
 from . import gff3
-# from . import occ
+from . import occ
